# Cache the partition pruning result in FilteredDeltaScanImpl to improve performance