--- a/standalone/src/test/scala/io/delta/standalone/internal/BenchmarkPartitionFilterRecordCachingSuite.scala
+++ b/standalone/src/test/scala/io/delta/standalone/internal/BenchmarkPartitionFilterRecordCachingSuite.scala
@@ -1,3 +1,19 @@
+/*
+ * Copyright (2020-present) The Delta Lake Project Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.delta.standalone.internal
 
 import java.io.File
@@ -50,7 +66,7 @@ class BenchmarkPartitionFilterRecordCachingSuite extends FunSuite with Logging {
       "col1" -> (i % 2).toString,
       "col2" -> (i % 3).toString,
       "col3" -> (i % 2).toString,
-      "col4" -> (i % 5).toString,
+      "col4" -> (i % 5).toString
     )
     AddFile(i.toString, partitionValues, 1L, 1L, dataChange = true)
   }
@@ -62,7 +78,7 @@ class BenchmarkPartitionFilterRecordCachingSuite extends FunSuite with Logging {
     ),
     new And(
       new EqualTo(partitionSchema.column("col3"), Literal.of("1")),
-      new EqualTo(partitionSchema.column("col4"), Literal.of("4")),
+      new EqualTo(partitionSchema.column("col4"), Literal.of("4"))
     )
   )