Cache the partition pruning result in FilteredDeltaScanImpl to improve performance (#399)
Signed-off-by: sonhmai <14060682+sonhmai@users.noreply.github.com>

Benchmark result: elapsed times, total 200 scans, 10k AddFile ops
- With Caching: total 6.448519193 secs, avg per scan 32.242595965 ms
- Without Caching: 7.452229266 secs, avg per run 37.26114633 ms
- Caching saves on avg 5.018550365 ms per scan