@scottsand-db 2022-08-10
@sonhmai - looking good! Could you add a micro benchmark? with perhaps a few layers of partitioning?