@sonhmai 2022-08-11
> @sonhmai - looking good! Could you add a micro benchmark? with perhaps a few layers of partitioning?

sure will do this. just pushed the other first