@sonhmai 2022-08-19
> LGTM! @sonhmai before I merge, can you please add the results of the micro benchmark to the PR description?

@scottsand-db Added to the first comment. If that is not the right place, pls let me know.